requirements.txt
# Candidate 27 and Candidate 16

numpy
matplotlib
seaborn
pandas
torch
torchvision
Pillow
albumentations
opencv-python-headless
tqdm # only used in preprocess_images.py
